Move AccountMapper to separate class
