+++ src/main/java/com/diablo3CharViewer/Main.java
@@ -4,6 +4,7 @@ import com.diablo3CharViewer.api_handlers.AccountHandlerApi;
 import com.diablo3CharViewer.api_handlers.HeroHandlerApi;
 import com.diablo3CharViewer.api_handlers.ItemHandlerApi;
 import com.diablo3CharViewer.data_models.AccountDataModel;
+import com.diablo3CharViewer.json_mappers.AccountMapper;
 import com.diablo3CharViewer.token.FetchToken;
 import com.diablo3CharViewer.token.Token;
 
@@ -13,7 +14,8 @@ public class Main {
     public static void main(String[] args) {
 
         FetchToken fetchToken = new FetchToken();
-        AccountHandlerApi accountHandlerApi = new AccountHandlerApi();
+        //AccountHandlerApi accountHandlerApi = new AccountHandlerApi();
+        AccountMapper accountMapper = new AccountMapper();
         HeroHandlerApi heroHandlerApi = new HeroHandlerApi();
         ItemHandlerApi itemHandlerApi = new ItemHandlerApi(); //dac tworzenie klas po podaniu battleTag lub item Id z podaną wartoscia dla tego pola
         CharacterViewerManager characterViewerManager = new CharacterViewerManager();
@@ -30,7 +32,7 @@ public class Main {
 
             switch (wybor) {
                 case "1":
-                    characterViewerManager.showProfile(scanner, accountHandlerApi, fetchToken);
+                    characterViewerManager.showProfile(scanner, accountMapper, fetchToken);
                     break;
                 case "2":
                     characterViewerManager.showHero(scanner, heroHandlerApi, fetchToken);
